dcmstan: Generate 'Stan' Code for Estimating Diagnostic Classification Models

Diagnostic classification models are psychometric models used to categorically estimate respondents mastery, or proficiency, on a set of predefined skills (Bradshaw, 2016, doi:10.1002/9781118956588.ch13 ). Diagnostic models can be estimated with 'Stan'; however, the necessary scripts can be long and complicated. This package automates the creation of 'Stan' scripts for diagnostic classification models. Specify different types of diagnostic models, define prior distributions, and automatically generate the necessary 'Stan' code for estimating the model.

See also

Author

Maintainer: W. Jake Thompson wjakethompson@gmail.com (ORCID)

Authors:

Jeffrey Hoover jeffrey.c.hoover@gmail.com (ORCID)

Auburn Jimenez auburn.jimenez34@gmail.com (ORCID)

Other contributors:

University of Kansas [copyright holder]

Institute of Education Sciences [funder]